Rentenreform/requirements.txt
pyyaml>=5.4
numpy>=1.20
